pyproject.toml
[project]
name = "giga"
version = "1.0.0"
description = "Codebase of GIGA project."
readme = "README.md"
authors = [
    { name = "Anton Zubekhin", email = "azubekhi@mpi-inf.mpg.de"}
]
requires-python = ">=3.12, <3.14"
dependencies = [
    "jaxtyping>=0.2.36",
    "loguru>=0.7.3",
    "numpy>=2.1.3",
    "platformdirs>=4.3.6",
    "pillow>=11.1.0",
    "typer>=0.15.1",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "setuptools==80.9.0",
    "einops==0.8.1",
    "timm==1.0.15",
    "omegaconf>=2.3.0",
    "torchmetrics>=1.7.1",
    "lightning>=2.5.3",
    "h5py>=3.13.0",
    "wandb>=0.19.11",
    "opt-einsum>=3.4.0",
    "opencv-python>=4.11.0.86",
    "huggingface-hub>=0.32.3",
    "nvdiffrast",
    "gsplat",
    "gdown>=5.2.0",
]


[tool.uv]
environments = ["sys_platform == 'linux'", "sys_platform == 'win32'"]

[tool.uv.extra-build-dependencies]
gsplat = ["setuptools==80.9.0", "torch==2.8.0"]

[build-system]
requires = ["setuptools==80.9.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["giga*"]
namespaces = true

[tool.ruff]
line-length = 140
lint.ignore = ["E741", "E743", "F722"] # ignore ambiguous variable names
lint.extend-select = ["I"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403"] # exclude imports from __init__.py

[tool.uv.sources]
torch = [{ index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" }]
torchvision = [{ index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" }]
nvdiffrast = { git = "https://github.com/NVlabs/nvdiffrast.git", tag = "v0.3.3" }
gsplat = { git = "https://github.com/nerfstudio-project/gsplat.git", tag = "v1.5.3" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
